cleanup: move changePassword && rename controllers
