+++ backend/src/controllers/auth/password/password.routes.ts
@@ -1,15 +1,15 @@
 import express from "express";
 
 import { authMiddleware } from "../../../middlewares/auth.middleware";
-import { changePassword } from "../../users/changePassword/changePassword.controller";
 
-import { forgotPassword } from "./forgotPassword/forgotPassword.controller";
+import { changePasswordController } from "./changePassword/changePassword.controller";
+import { forgotPasswordController } from "./forgotPassword/forgotPassword.controller";
 import { resetPasswordController } from "./resetPassword/resetPassword.controller";
 
 export const passwordRoutes = express.Router()
 
-passwordRoutes.post("/forgot", forgotPassword)
+passwordRoutes.post("/forgot", forgotPasswordController)
 passwordRoutes.post("/reset", resetPasswordController)
-passwordRoutes.post("/change", authMiddleware, changePassword)
+passwordRoutes.post("/change", authMiddleware, changePasswordController)
 
 export default passwordRoutes
